+++ packages/lexical-playground/src/utils/emoji-list.ts
